LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/container-queries/anchored-fallback-try-tactic-expected.txt


FAIL @container anchored(fallback) matching flip-block assert_equals: expected "yes" but got "no"
FAIL @container anchored(fallback) matching flip-inline flip-block assert_equals: expected "yes" but got "no"
